pyproject.toml
[project]
name = "projeto-dados"
version = "0.1.0"
description = ""
authors = [
    {name = "Kauã D. Rocha"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openpyxl (==3.1.5)",
    "pandas (==2.2.3)",
    "NumPy (==2.2.0)",
    "matplotlib (==3.10.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
